requirements.txt
Django==4.2.11
djangorestframework==3.15.2
django-cors-headers==4.3.1
requests==2.31.0
